[종원] 채팅방 별 알림 온오프 추가 외 - 프로필에 한 줄 소개 디자인 추가 - 내가 쓴 글 등 3개의 배너 디자인 변경 - 매너 그룹 int형 타입에러 수정 - 버전 확인이 더 자주되게 변경 - 작성자 프로필 디자인 변경 - 채팅 오류 수정

+++ lib/resources/resources.dart
@@ -55,7 +55,7 @@ var CategoryList = List.of([
   "쇼핑",
 ]);
 
-Color getColorForScore(int score) {
+Color getColorForScore(double score) {
   if (score < 20) {
     return Colors.redAccent;
   } else if (score < 40) {
